CloudFunctions/premonotion/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.21.5
google-cloud-aiplatform